Fix bug which can't delete legacy rules and groups
Delete the legacy rule and group at last when updating security policy.
Split large file.
Cherry-pick from commits of v4.0.0 branch.
