+++ pkg/nsx/services/parse_test.go
@@ -1,0 +1,11 @@
+package services
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGetCluster(t *testing.T) {
+	assert.Equal(t, "k8scl-one", getCluster(service))
+}
