+++ pkg/nsx/services/store.go
@@ -7,9 +7,11 @@ import (
 	"sync"
 
 	"github.com/vmware/vsphere-automation-sdk-go/lib/vapi/std/errors"
-	"github.com/vmware/vsphere-automation-sdk-go/runtime/bindings"
 	"github.com/vmware/vsphere-automation-sdk-go/services/nsxt/model"
+	"k8s.io/client-go/tools/cache"
 
+	"github.com/vmware-tanzu/nsx-operator/pkg/config"
+	"github.com/vmware-tanzu/nsx-operator/pkg/nsx"
 	util2 "github.com/vmware-tanzu/nsx-operator/pkg/nsx/util"
 	"github.com/vmware-tanzu/nsx-operator/pkg/util"
 )
@@ -74,7 +76,8 @@ func keyFunc(obj interface{}) (string, error) {
 
 func queryTagCondition(service *SecurityPolicyService) string {
 	return fmt.Sprintf("tags.scope:%s AND tags.tag:%s",
-		strings.Replace(util.TagScopeCluster, "/", "\\/", -1), strings.Replace(service.NSXClient.NsxConfig.Cluster, ":", "\\:", -1))
+		strings.Replace(util.TagScopeCluster, "/", "\\/", -1),
+		strings.Replace(service.NSXClient.NsxConfig.Cluster, ":", "\\:", -1))
 }
 
 func queryGroup(service *SecurityPolicyService, wg *sync.WaitGroup, fatalErrors chan error) {
@@ -92,16 +95,18 @@ func queryGroup(service *SecurityPolicyService, wg *sync.WaitGroup, fatalErrors 
 		if err != nil {
 			fatalErrors <- err
 		}
-		typeConverter := service.NSXClient.RestConnector.TypeConverter()
 		for _, g := range response.Results {
-			a, err := typeConverter.ConvertToGolang(g, model.GroupBindingType())
+			a, err := Converter.ConvertToGolang(g, model.GroupBindingType())
 			if err != nil {
 				for _, e := range err {
 					fatalErrors <- e
 				}
 			}
 			c, _ := a.(model.Group)
-			service.GroupStore.Add(c)
+			err2 := service.GroupStore.Add(c)
+			if err2 != nil {
+				fatalErrors <- err2
+			}
 		}
 		cursor = response.Cursor
 		if cursor == nil {
@@ -129,16 +134,18 @@ func querySecurityPolicy(service *SecurityPolicyService, wg *sync.WaitGroup, fat
 		if err != nil {
 			fatalErrors <- err
 		}
-		typeConverter := service.NSXClient.RestConnector.TypeConverter()
 		for _, g := range response.Results {
-			a, err := typeConverter.ConvertToGolang(g, model.SecurityPolicyBindingType())
+			a, err := Converter.ConvertToGolang(g, model.SecurityPolicyBindingType())
 			if err != nil {
 				for _, e := range err {
 					fatalErrors <- e
 				}
 			}
 			c, _ := a.(model.SecurityPolicy)
-			service.SecurityPolicyStore.Add(c)
+			err2 := service.SecurityPolicyStore.Add(c)
+			if err2 != nil {
+				fatalErrors <- err2
+			}
 		}
 		cursor = response.Cursor
 		if cursor == nil {
@@ -166,16 +173,18 @@ func queryRule(service *SecurityPolicyService, wg *sync.WaitGroup, fatalErrors c
 		if err != nil {
 			fatalErrors <- err
 		}
-		typeConverter := service.NSXClient.RestConnector.TypeConverter()
 		for _, g := range response.Results {
-			a, err := typeConverter.ConvertToGolang(g, model.RuleBindingType())
+			a, err := Converter.ConvertToGolang(g, model.RuleBindingType())
 			if err != nil {
 				for _, e := range err {
 					fatalErrors <- e
 				}
 			}
 			c, _ := a.(model.Rule)
-			service.RuleStore.Add(c)
+			err2 := service.RuleStore.Add(c)
+			if err2 != nil {
+				fatalErrors <- err2
+			}
 		}
 		cursor = response.Cursor
 		if cursor == nil {
@@ -196,15 +205,13 @@ func decrementPageSize(pageSize *int64) {
 }
 
 func transError(err error) error {
-	var typeConverter = bindings.NewTypeConverter()
-	typeConverter.SetMode(bindings.REST)
 	switch err.(type) {
 	case errors.ServiceUnavailable:
 		vapiError, _ := err.(errors.ServiceUnavailable)
 		if vapiError.Data == nil {
 			return err
 		}
-		data, errs := typeConverter.ConvertToGolang(vapiError.Data, model.ApiErrorBindingType())
+		data, errs := Converter.ConvertToGolang(vapiError.Data, model.ApiErrorBindingType())
 		if len(errs) > 0 {
 			return err
 		}
@@ -216,4 +223,95 @@ func transError(err error) error {
 		return err
 	}
 	return err
+}
+
+func (service *SecurityPolicyService) OperateSecurityStore(sp *model.SecurityPolicy) error {
+	if sp == nil {
+		return nil
+	}
+	if sp.MarkedForDelete != nil && *sp.MarkedForDelete {
+		err := service.SecurityPolicyStore.Delete(*sp) // Pass in the object to be deleted, not the pointer
+		log.V(1).Info("delete security policy from store", "securitypolicy", sp)
+		if err != nil {
+			return err
+		}
+	} else {
+		err := service.SecurityPolicyStore.Add(*sp)
+		log.V(1).Info("add security policy to store", "securitypolicy", sp)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (service *SecurityPolicyService) OperateRuleStore(sp *model.SecurityPolicy) error {
+	for _, rule := range sp.Rules {
+		if rule.MarkedForDelete != nil && *rule.MarkedForDelete {
+			err := service.RuleStore.Delete(rule)
+			log.V(1).Info("delete rule from store", "rule", rule)
+			if err != nil {
+				return err
+			}
+		} else {
+			err := service.RuleStore.Add(rule)
+			log.V(1).Info("add rule to store", "rule", rule)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func (service *SecurityPolicyService) OperateGroupStore(gs *[]model.Group) error {
+	for _, group := range *gs {
+		if group.MarkedForDelete != nil && *group.MarkedForDelete {
+			err := service.GroupStore.Delete(group)
+			log.V(1).Info("delete group from store", "group", group)
+			if err != nil {
+				return err
+			}
+		} else {
+			err := service.GroupStore.Add(group)
+			log.V(1).Info("add group to store", "group", group)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+// InitializeSecurityPolicy sync NSX resources
+func InitializeSecurityPolicy(NSXClient *nsx.Client, cf *config.NSXOperatorConfig) (*SecurityPolicyService, error) {
+	wg := sync.WaitGroup{}
+	wgDone := make(chan bool)
+	fatalErrors := make(chan error)
+
+	wg.Add(3)
+	service := &SecurityPolicyService{NSXClient: NSXClient}
+	service.GroupStore = cache.NewIndexer(keyFunc, cache.Indexers{util.TagScopeNamespace: namespaceIndexFunc, util.TagScopeSecurityPolicyCRUID: securityPolicyCRUIDScopeIndexFunc})
+	service.SecurityPolicyStore = cache.NewIndexer(keyFunc, cache.Indexers{util.TagScopeSecurityPolicyCRUID: securityPolicyCRUIDScopeIndexFunc})
+	service.RuleStore = cache.NewIndexer(keyFunc, cache.Indexers{util.TagScopeSecurityPolicyCRUID: securityPolicyCRUIDScopeIndexFunc})
+	service.NSXConfig = cf
+
+	go queryGroup(service, &wg, fatalErrors)
+	go querySecurityPolicy(service, &wg, fatalErrors)
+	go queryRule(service, &wg, fatalErrors)
+	go func() {
+
+		wg.Wait()
+		close(wgDone)
+	}()
+
+	select {
+	case <-wgDone:
+		break
+	case err := <-fatalErrors:
+		close(fatalErrors)
+		return service, err
+	}
+
+	return service, nil
 }
